requirements.txt
# Python dependencies
python-dotenv
psycopg2-binary
telethon
dotenv
pandas
dbt-postgres
